+++ changeLog.md
@@ -1,18 +1,27 @@
 # ChangeLog
 
 ### 0.0.1
+
 [x] Add Change Log
 [x] Make app stabile
 [x] Add Change Convertor file
 
 ### 0.0.2 [working to]
+[x] Moved table page
+[x] Expanded functionality of the table page with helpers
+[ ] Add Table header title retrieve
 [ ] Add Documentation for all possible props
 
 ### 0.1.0 [milestone 1]
+
+[ ] Add borrowed tool functionality
+[ ] Add barcode functionality
 [ ] Add all tools
 
-### 0.2.0 
+### 0.2.0
+
 [ ] Add CI && git logic for release
 
 ### 1.0.0 [milestone 2]
+
 [ ] Add a real release